requirements_dev.txt
torch
paddleocr 
paddlepaddle
PyMuPDF
streamlit
pandas
pyyaml
wget
dvc
xlsxwriter
tox
pytest
flake8
black
-e .